New representation of polyhedral structures
Three pieces of data:

* a cardinality, N, whose interpretation is that the set {0,1,...,N-1}
are the polytopes;

* a "type vector" v, a vector of length N with entries among {0,1,2}
which is interpreted as a specification of the "types" of the elements
{0,1,...,N-1}: v[j] = 0 means that j is a vertex, v[j] = 1 means that
j is an edge, and v[j] = 2 means that j is a face.

* an incidence matrix (same as before).

There was a bug before, in the previous representation of individuals,
which led to nonsense in certain cases.  In any case, a more compact
representation of polyhedron structures was wanted, and working on
that new representation fixed the bug.  (The old bug was such that
crossover on two individuals could lead to nonsense, such as two
polyhedron of cardinality 10 being combined to give a polyhedron of
cardinality different from 10, which violates the homogeneity
condition that is being taken for granted.)

To combine these new individuals, we now do the usual GA crossover of
two one-dimensional arrays.