Don't generate "typed" polyhedron structures in the initial population.
Now that the condition for non-emptyness of the domains is one of the
axioms to be satisfied and thus contributes to the fitness of
individuals, I'll let selection try to solve this constraint, rather
than forcing it to be satisfied in the initial generation.  Besides,
crossover can take individuals that satisfy the non-emptyness
conditions and produce an individual that fails to satisfy them, so
what difference does it really make that the initial generation
consists entirely of the right kind of individuals?