Weighted fitness function.
I really have no idea how to appropriately weigh the axioms.

One trick I noticed, when seeing many searches terminate with a best
individual consisting only of vertices, or only of faces, was to take
the conjunction of the axioms asserting non-emptyness of the types,
and assign a high weight to that conjunction.  It didn't help nearly
as much as I thought it would. So odd.